Create a navbar search
Create a navbar bootstrap styled with a form search.
Create two functions to handle the submit and the change of the input.
Use the function handleSubmit for pass data to parent componen
Create two functions to handle the submit and the change of the input.
Use the function handleSubmit for pass data to parent component.