import NavbarSearch component
Use NavbarSearch child component to trigger the function that allows
ContentBox to search a specific pokemon.
Use render conditionals for display the data default or the data of the
search.
Add the state pokemon to store the data of an specific pokemon.
Add the function handleSubmit to be triggered by the child component
props.